refactor(search): share query normalizer + drop fixture-tuned stop-tokens (#1572, #1573)
#1572: normalizeSearchText was duplicated byte-for-byte in
src/shared/action_handlers/entity_handlers.ts and as normalizeConceptPhrase in
src/services/schema_registry.ts — the registry kept a private copy because
entity_handlers already imports schema_registry, so importing back would cycle.
A future tokenization change (Unicode folding, stemming) would have had to be
mirrored or the concept->type synonym bridge would silently stop matching.
Extract a dependency-free src/shared/search_normalization.ts imported by both;
entity_handlers re-exports the name for back-compat.

#1573: PARTIAL_MATCH_STOP_TOKENS mixed generic English function words with
site-/fixture-specific noise tokens (now / try / site) tuned to suppress one
decoy in the #1551 test. That is hardcoded query-domain knowledge in a generic
module. Drop the three noise tokens; precision on descriptive queries is the
job of PARTIAL_MATCH_MIN_OVERLAP_RATIO (0.5), not an ad-hoc denylist that would
accrete an entry every time a query trips the gate. No test depended on the
removed tokens (verified: the lexical_search #1551 precision cases still pass
via the overlap threshold).

Tests: new search_normalization unit suite (6, incl. asserting the
entity_handlers re-export is the same function so the two consumers can't
drift); lexical_search + entity_search_mode + schema-registry suites green;
type-check / lint / format clean. No API or behavior change.
